move audit into its own package
