+++ service/internal/auth/authn.go
@@ -18,7 +18,7 @@ import (
 	"github.com/lestrrat-go/jwx/v2/jwk"
 	"github.com/lestrrat-go/jwx/v2/jws"
 	"github.com/lestrrat-go/jwx/v2/jwt"
-	"github.com/opentdf/platform/service/internal/logger"
+	"github.com/opentdf/platform/service/internal/logger/audit"
 	"google.golang.org/grpc"
 	"google.golang.org/grpc/codes"
 	"google.golang.org/grpc/metadata"
@@ -308,7 +308,7 @@ func (a Authentication) checkToken(ctx context.Context, authHeader []string, dpo
 
 	// Get actor ID (sub) from unverified token for audit and add to context
 	actorID := unverifiedToken.Subject()
-	ctx = context.WithValue(ctx, logger.ActorIdContextKey, actorID)
+	ctx = context.WithValue(ctx, audit.ActorIdContextKey, actorID)
 
 	// Get issuer from unverified token
 	issuer := unverifiedToken.Issuer()
