+++ service/internal/logger/audit/auditPolicyUtils.go
@@ -1,4 +1,4 @@
-package logger
+package audit
 
 import (
 	"context"
